+++ pkg/providers/appsec/testdata/ase/modules/security/url-protection-action.tf
@@ -3,7 +3,7 @@
 resource "akamai_appsec_url_protection_action" "default_policy_50411" {
   config_id                 = local.config_id
   security_policy_id        = akamai_appsec_security_policy.default_policy.security_policy_id
-  url_protection_rule_id    = 50411
+  url_protection_policy_id  = 50411
   max_rate_threshold_action = "alert"
   load_shedding_action      = "deny"
 }
@@ -11,7 +11,7 @@ resource "akamai_appsec_url_protection_action" "default_policy_50411" {
 resource "akamai_appsec_url_protection_action" "default_policy_50412" {
   config_id                 = local.config_id
   security_policy_id        = akamai_appsec_security_policy.default_policy.security_policy_id
-  url_protection_rule_id    = 50412
+  url_protection_policy_id  = 50412
   max_rate_threshold_action = "deny"
   load_shedding_action      = "none"
 }